refactor(devpulse): the watchdog module outgrew the ceiling its own three rewrites pushed it through - the presentation layer moves out, and the first attempt at moving it was wrong. apps/modules/watchdog.py stood at 692 lines against a hard 600-line module limit, which the CI standards gate scores as a failure: too large for AI comprehension. It did not get there by accident - r2 split detection from delivery, r3 gated the polling, r4 deleted the daemon and added the register, and each round left its output rendering in the module. So the cut is not at an arbitrary line number: the display functions were the cohesive third that shared no state with the routing code and were already only called at the end of a subcommand. presenter.py now owns the result printers, the per-watch-type tail renderers with their lookup table, the status-line formatter, the dispatch-lag line and the kill-result line. The module drops to 519. THE FIRST SPLIT FAILED THE AUDIT IN A NEW WAY, which is the part worth recording: a handler may not call console.print() - that is the cli standard's separation rule, modules do CLI and handlers compute - and every handler must carry json_handler plus a log_operation. So a green module ceiling bought two fresh violations on the file I had just created. The house already answers this: apps/handlers/feedback/compose.py and inbox.py are display handlers carrying a written cli bypass whose reason is that console.print IS the output. presenter.py now follows that precedent exactly rather than inventing a fourth answer, with a bypass reason naming why display is its purpose and where it came from. The log_operation is NOT ceremony added to satisfy a checker: it records each read of the dispatch register, and since r4 made looking itself the crash detector - an entry past expected_by becomes late whether or not anything is running - that record is the difference between a week in which nothing was overdue and a week in which nobody looked. It is FIFO-bounded at 100 entries, so it cannot become the 1277-of-1279-identical-lines problem c627cd7e fixed in drone's log. Two tests that pinned _format_status_line and _print_kill_result on the module were MOVED to the presenter rather than re-exported through an alias: a test should name where the code actually lives, and an alias kept only to keep a test passing is a line of code with no other reason to exist. Verified: devpulse 99% to 100% on every standard including modules, 593 tests green, and the live routed command re-run after the refactor - drone @devpulse watchdog status rendered the wire row and the dispatch line, exercising both moved render paths. That live check is deliberate: last night I hot-edited a routed module while an agent was calling it and left it answering rc=1 with zero bytes for twenty seconds, and a passing unit suite would not have caught it.
