+++ src/aipass/devpulse/apps/modules/watchdog.py
@@ -33,6 +33,7 @@ from typing import List
 from aipass.prax.apps.modules.logger import system_logger as logger
 from aipass.cli.apps.modules import console, err_console, error
 from aipass.devpulse.apps.handlers.json import json_handler
+from aipass.devpulse.apps.handlers.watchdog import presenter
 
 _VALID_SUBCOMMANDS = ["agent", "baseline", "timer", "schedule", "status", "cancel", "list"]
 _DEFAULT_AGENT_TIMEOUT = 600
@@ -269,7 +270,7 @@ def _handle_timer(sub_args: List[str]) -> bool:
             error("Usage: watchdog timer start <name>")
             return True
         result = timer_mod.timer_start(sub_args[1])
-        _print_timer_result(result)
+        presenter.print_timer_result(result)
         return True
 
     if action == "stop":
@@ -277,12 +278,12 @@ def _handle_timer(sub_args: List[str]) -> bool:
             error("Usage: watchdog timer stop <name>")
             return True
         result = timer_mod.timer_stop(sub_args[1])
-        _print_timer_result(result)
+        presenter.print_timer_result(result)
         return True
 
     if action == "list":
         snapshot = timer_mod.timer_list()
-        _print_timer_list(snapshot)
+        presenter.print_timer_list(snapshot)
         return True
 
     if action == "report":
@@ -296,7 +297,7 @@ def _handle_timer(sub_args: List[str]) -> bool:
         logger.warning("[watchdog] invalid timer duration %r: %s", action, exc)
         error(f"Invalid duration: {action} ({exc})")
         return True
-    _print_timer_result(result)
+    presenter.print_timer_result(result)
     return True
 
 
@@ -327,63 +328,8 @@ def _handle_schedule(sub_args: List[str]) -> bool:
         error(f"Invalid schedule: {time_str} ({exc})")
         return True
 
-    _print_schedule_result(result)
+    presenter.print_schedule_result(result)
     return True
-
-
-def _print_schedule_result(result: dict) -> None:
-    """Render a schedule handler return dict as CLI output."""
-    scheduled_for = result.get("scheduled_for", "?")
-    elapsed = result.get("elapsed", 0)
-    console.print(f"[bold]watchdog schedule[/bold] woke after {elapsed}s (scheduled_for={scheduled_for})")
-    if result.get("command"):
-        exit_code = result.get("command_exit_code")
-        console.print(f"  command: {result['command']} -> exit={exit_code}")
-        stdout = result.get("command_stdout") or ""
-        stderr = result.get("command_stderr") or ""
-        if stdout:
-            console.print(f"  stdout: {stdout.rstrip()}")
-        if stderr:
-            console.print(f"  stderr: {stderr.rstrip()}")
-
-
-def _print_timer_result(result: dict) -> None:
-    """Render a timer handler return dict as a single CLI line."""
-    state = result.get("state", "unknown")
-    name = result.get("name") or result.get("duration") or ""
-    if state == "error":
-        error(f"timer {name}: {result.get('reason', 'unknown error')}")
-        return
-    if state == "stopped":
-        console.print(f"[bold]timer[/bold] {name} stopped -> elapsed={result.get('human', '?')}")
-        return
-    if state == "started":
-        console.print(f"[bold]timer[/bold] {name} started at {result.get('started_at', '?')}")
-        return
-    if state == "woke":
-        console.print(f"[bold]timer[/bold] {name} woke after {result.get('elapsed', 0)}s")
-        return
-    console.print(f"[dim]timer result:[/dim] {result}")
-
-
-def _print_timer_list(snapshot: dict) -> None:
-    """Pretty-print the ``timer_list`` snapshot."""
-    active = snapshot.get("active", [])
-    history = snapshot.get("history", [])
-    console.print("[bold]Active timers:[/bold]")
-    if active:
-        for item in active:
-            console.print(f"  - {item['name']}  elapsed {item['human']}  (started {item.get('started_at', '?')})")
-    else:
-        console.print("  (none)")
-    console.print("[bold]History:[/bold]")
-    if history:
-        for item in history:
-            console.print(
-                f"  - {item['name']}  {item['human']}  ({item.get('started_at', '?')} -> {item.get('stopped_at', '?')})"
-            )
-    else:
-        console.print("  (none)")
 
 
 def _handle_agent(sub_args: List[str]) -> bool:
@@ -504,78 +450,6 @@ def _load_timer_module_for_format():
     return importlib.import_module("aipass.devpulse.apps.handlers.watchdog.timer")
 
 
-def _tail_agent(meta: dict) -> str:
-    return f"{meta.get('agent_id', '?')} (timeout={meta.get('timeout_seconds', '?')}s)"
-
-
-def _tail_baseline(meta: dict) -> str:
-    """A pre-r4 detection daemon, if one is somehow still standing.
-
-    r4 deleted the daemon, so this row should never appear again — which is
-    exactly why the renderer stays. A row nothing can render is a row nobody
-    sees, and an orphaned daemon from an older binary is precisely the thing
-    the operator needs shown to them.
-    """
-    role = meta.get("role") or "legacy"
-    return f"role={role} scope={meta.get('scope', '?')} (tick={meta.get('tick_seconds', '?')}s)"
-
-
-def _tail_baseline_wire(meta: dict) -> str:
-    """The wire. ``session`` is the whole of the health question.
-
-    This used to print ``daemon_pid`` — a field r4's wire no longer writes, so
-    it rendered as a permanent ``daemon_pid=?`` that read like a fault. The
-    first replacement was ``tasks_dir``, which was the same mistake wearing a
-    different name: a Monitor child's stdout is a socket and has no tasks dir,
-    so the healthy case printed ``tasks=?`` forever. A field that reads ``?``
-    on the happy path teaches the operator to ignore the row.
-
-    ``via`` is the honest one. Only ``monitor`` means covered — ``background``
-    is the wrapper that notifies on exit only (a continuous wire never exits,
-    so zero wakes), and ``foreground`` means nobody is listening at all.
-    """
-    return f"session={meta.get('session') or 'NONE (fg)'} via={meta.get('wrapper') or 'unrecorded'}"
-
-
-def _tail_timer(meta: dict) -> str:
-    return f"duration={meta.get('duration', '?')}"
-
-
-def _tail_schedule(meta: dict) -> str:
-    cmd = meta.get("command")
-    cmd_repr = f' cmd="{cmd}"' if cmd else ""
-    return f"scheduled={meta.get('scheduled_for', '?')}{cmd_repr}"
-
-
-# Type -> tail renderer. A table rather than an if/elif chain: the chain nested
-# one level deeper per watch type, so adding a seventh kind of watch was a
-# standards failure on arrival.
-_TAIL_BY_TYPE = {
-    "agent": _tail_agent,
-    "baseline": _tail_baseline,
-    "baseline_wire": _tail_baseline_wire,
-    "timer": _tail_timer,
-    "schedule": _tail_schedule,
-}
-
-
-def _format_status_line(watch: dict, format_human) -> str:
-    """One-line renderer for a single watch entry in the status output."""
-    handle = watch.get("handle", "?")
-    wtype = watch.get("type", "?")
-    elapsed = int(watch.get("elapsed_seconds", 0))
-    pid = watch.get("pid", "?")
-    meta = watch.get("metadata") or {}
-
-    # An unknown type still renders — raw metadata beats hiding the row, because
-    # a watch nobody can see is a watch nobody retires.
-    render = _TAIL_BY_TYPE.get(wtype)
-    tail = render(meta) if render is not None else str(meta)
-
-    # Escape the [ so Rich console doesn't interpret it as a style tag.
-    return f"  \\[{handle}]  {wtype:<8}  {format_human(elapsed):<10}  pid={pid}  {tail}"
-
-
 def _handle_status() -> bool:
     """Read the watch registry, prune stale entries, pretty-print active watches."""
     registry_mod = _load_registry_module()
@@ -599,52 +473,15 @@ def _handle_status() -> bool:
     console.print(f"{len(post)} active watch(es):")
     console.print()
     for watch in post:
-        console.print(_format_status_line(watch, timer_mod.format_human))
+        console.print(presenter.format_status_line(watch, timer_mod.format_human))
 
     if pruned:
         console.print(f"[dim]Pruned {pruned} stale watch(es).[/dim]")
     else:
         console.print("[dim]No stale watches to prune.[/dim]")
 
-    _print_delivery_lag()
+    presenter.print_delivery_lag()
     return True
-
-
-def _print_delivery_lag() -> None:
-    """One line of dispatch truth: what this seat has out, and what is late.
-
-    r4 replaced the old observable (events written vs bytes delivered) because
-    both files it read are gone with the daemon. This reads the REGISTER
-    instead, and it is the whole of the crash coverage: an entry past its
-    expected_by is late whether or not anything is running, so simply looking
-    is the detector. Nothing polls for this — it becomes true on its own and is
-    noticed by whoever next asks.
-
-    A missing register is reported as exactly that, never as "none outstanding".
-    """
-    dispatches_mod = importlib.import_module("aipass.devpulse.apps.handlers.watchdog.dispatches")
-    try:
-        open_now = dispatches_mod.outstanding()
-        late = dispatches_mod.overdue()
-    except dispatches_mod.RegisterUnavailable as exc:
-        # Printed AND logged: the print tells whoever ran `status` right now,
-        # the log is what remains if the crash coverage was blind for a week
-        # and nobody was watching the terminal when it went blind.
-        logger.warning("[watchdog] dispatch register unavailable: %s", exc)
-        console.print(f"[dim]Dispatch register unavailable: {exc}[/dim]")
-        return
-
-    if late:
-        names = ", ".join(str(e.get("target", "?")) for e in late)
-        console.print(
-            f"[yellow]Dispatches overdue: {len(late)}[/yellow] ({names}) — "
-            f"{len(open_now)} outstanding. An overdue entry never reported back."
-        )
-    elif open_now:
-        names = ", ".join(str(e.get("target", "?")) for e in open_now)
-        console.print(f"[dim]Dispatches outstanding: {len(open_now)} ({names}), none overdue.[/dim]")
-    else:
-        console.print("[dim]Dispatches: none outstanding.[/dim]")
 
 
 def _handle_list() -> bool:
@@ -654,16 +491,6 @@ def _handle_list() -> bool:
     differentiating wasn't worth the divergence.
     """
     return _handle_status()
-
-
-def _print_kill_result(result: dict) -> None:
-    """Render a single ``registry.kill_watch`` result on one line."""
-    handle = result.get("handle", "?")
-    killed = result.get("killed", False)
-    was_alive = result.get("was_alive", False)
-    reason = result.get("reason", "")
-    status = "KILLED" if killed else "FAILED"
-    console.print(f"  \\[{handle}] {status} was_alive={was_alive} reason={reason}")
 
 
 def _handle_cancel(sub_args: List[str]) -> bool:
@@ -681,12 +508,12 @@ def _handle_cancel(sub_args: List[str]) -> bool:
             return True
         console.print(f"[bold]Cancelling {len(results)} watch(es):[/bold]")
         for result in results:
-            _print_kill_result(result)
+            presenter.print_kill_result(result)
         return True
 
     handle = sub_args[0]
     result = registry_mod.kill_watch(handle)
-    _print_kill_result(result)
+    presenter.print_kill_result(result)
     if not result.get("killed", False):
         logger.info("[watchdog] cancel failed handle=%s", handle)
     return True
